[CPP][Analyzers]Turn on warning 4244 (#22019)
* Turn on warning 4244

fix one occurrence of 4100

* changes from comments

* toupper on wchar

+++ src/modules/ShortcutGuide/ShortcutGuide/overlay_window.cpp
@@ -619,7 +619,7 @@ void D2DOverlayWindow::render(ID2D1DeviceContext5* d2d_dc)
     d2d_dc->Clear();
     int x_offset = 0, y_offset = 0;
     auto current_anim_value = (float)animation.value(Animation::AnimFunctions::LINEAR);
-    SetLayeredWindowAttributes(hwnd, 0, (int)(255 * current_anim_value), LWA_ALPHA);
+    SetLayeredWindowAttributes(hwnd, 0, static_cast<byte>(255 * current_anim_value), LWA_ALPHA);
     double pos_anim_value = 1 - animation.value(Animation::AnimFunctions::EASE_OUT_EXPO);
     if (!tasklist_buttons.empty())
     {
